fun2learn_backend/requirements.txt
python-dotenv
fastapi
uvicorn
pymongo
sqlalchemy
psycopg2-binary
bcrypt
pyjwt